tests/wpt/web-platform-tests/fetch/sec-metadata/navigation.https.sub.html

<!DOCTYPE html>
<html>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/fetch/sec-metadata/resources/helper.js></script>
<script>
  test(t => {
    let expected = {
      "dest": "document",
      "mode": "navigate",
      "site": "none",
    };

    let actual = {
      "dest": "{{headers[sec-fetch-dest]}}",
      "mode": "{{headers[sec-fetch-mode]}}",
      "site": "{{headers[sec-fetch-site]}}",
      // Skipping `Sec-Fetch-User`, as the test harness isn't consistent here.
    };

    assert_header_equals(actual, expected);
  }, "This page's top-level navigation.");
</script>
